requirements.txt
tensorflow
numpy
pillow
matplotlib
streamlit
h5py